+++ CHANGELOG.md
@@ -18,6 +18,7 @@ New Rules
 Updated/Enhanced Rules and Bug Fixes
  - #777: Speed up **UnusedImport** rule ([Iulian Dragos](https://github.com/dragos)).
  - #762: **BooleanMethodReturnsNull** rule: Analyze contained closure expressions within their own context, not as part of outer method context.
+ - #781: **DuplicateNumberLiteral** rule: Fix *duplicateNumberMinimumValue* handling for Long and BigDecimal values.
  - #695: Refactor directory-handling in results. ([Ulrich Eckhardt](https://github.com/UlrichEckhardt))